normalizer/converters fail on empty content
These tools are tolerant of NEXUS content outside of
blocks. This content is skipped, but does not result
in an error. An unwanted side-effect of this was that
virtually any non-NEXUS file given to the NEXUS converter
would pass.
Now passing a file with no parsed content will result
in an error in these apps. No changes were made to the
library part of NCL, so no other clients need to change.