bug fix: making sure that the last branch anc is not equivalent.
If we say a source supports a node, then the most recent ancestor
of that node which is a branching ancestor must have a different
MRCA set (when the leaf set is pruned down to the set of the leaves
for the input tree).
This was not implemented correctly in previous versions.